+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Added the `send_precompiled_letter` method which allows the client to send letters as PDF files.
   * This requires two arguments - a reference for the letter and the PDF letter file. The file must conform to the Notify printing template.
-
+* Added support for document uploads using the `send_email` method.
 
 ## 2.8.0